feat: add dockerization and integrate redis
- added dockerfile
- added docker compose
- fixed bug nullable mapper

+++ my-app-core/src/main/java/by/kolp/myappcore/service/UserService.java
@@ -47,6 +47,9 @@ public class UserService {
         User user = findById(id)
                 .orElseThrow(() -> new NotFoundException(format("User not found with id %d", id)));
 
+        if(!user.getRole().equals("admin")) {
+            throw new BadRequestException(format("User %s is not admin", user));
+        }
         userRepository.deleteById(user.getId());
     }
